Refactor how queries that need validation are handled
With the ability to use a Query anywhere a Field was valid by using Query.asFunction, it became too difficult to manage when a root query would need to be validation for suspicious WHERE clauses. This new approach refactors SQL statement compilation to set an AtomicBoolean flag if any part of the Query being compiled requires validation. Before execution, SquidDatabase will check this flag. If it is set, it will compile again with the extra parentheses like it always has.

This refactor has the added benefit of cleaning up some of our SQL generation by bundling necessary builder objects (the SQL StringBuilder, the List<Objects> bound arguments, and the AtomicBoolean flag) into a single SqlBuilder class that can be passed to all SQL generating methods.