Make addPlugin() private
With the current plugin architecture, it doesn't make much sense for
outsiders to call this method. Clients can't really change the
PluginEnvironment, which means any code that runs will be part of a
plugin. Plugins are instantiated for each model spec, but addPlugin()
adds to the global list of plugins, so a plugin calling addPlugin()
would almost definitely be broken already.