Refactor Plugin API to be more like javax.annotation.processing.Processor.
Plugin is now an interface, with AbstractProcessor being a stubbed/empty implementation.
Plugins must provide a no-arg constructor to be instantiated, and will be initialized using
the init() method. If a plugin does not want to accept or process a given model spec,
it can simply return false in the init() method, and will be removed/ignored for
that model spec.

Javadoc cleanup

Code cleanup in ModelFileWriter and subclasses