Add nullability annotations to the public API of SquiDB
We use the jsr305 versions of the Nullable and Nonnull annotations,
because those are the versions that are best supported by j2objc.
Nullability annotations have been added to most of the public API,
although j2objc users should take note that since the JRE emul
project does not have nullability completeness, it is still best
to compile SquiDB and other shared Java code with j2objcc or to
use the -Wno-nullability-completeness flag for those files.

This commit does not include nullability annotations for
the generated getter and setter methods in model classes. That change
will be coming in a future enhancement.